Showtech sonic mgmt framework: Add handling for an output "status" field.
These changes make use of an additional output field for the "show
tech-support" command. The corresponding output-status leaf is added
to the set of output leaves in the Yang definition for the
sonic-show-techsupport module. This additional field is specified
in the SONiC Community HLD for the Management Framework "show tech-support"
feature, but was not present in the initial implementation of the feature.

The purpose of the additional field is to provide an indication of
"success" or detailed failure  status that is separate from the
output-filename field. With this change, the output-filename contains
either an output filename or a null string, and is no longer used to
describe errors in the event of a failure.