Delete commented out code from host_comm.go
This file contained commented out code to provide examples of synchronous and asynchronous invocation of the "HostQuery" and "HostQueryAsync" functions.

The examples are being deleted from this file to eliminate a block of commented out code. The equivalent information should, instead, be provided in the "SONiC_Host_Services_Developer_Guide" document.